+++ Lineaire algebra/Lineaire algebra/Vector.h
@@ -2,18 +2,18 @@
 class Vector
 {
 private: 
-	int direction; 
-	int length;
+	float direction; 
+	float length;
 	Vector* v;
 public:
-	Vector(int length, int direction); 
+	Vector(float length, float direction); 
 	~Vector();
-	void setdirection(int value); 
-	void setLength(int value);
-	int getlength(); 
-	int getdirection();
-	void addVectors(int richting, int lengte); 
-	Vector* scaleVector(int value); 
+	void setdirection(float value); 
+	void setLength(float value);
+	float getlength(); 
+	float getdirection();
+	void addVectors(float richting, float lengte); 
+	Vector* scaleVector(float value); 
 	Vector* combineVectors(Vector &v);
 	Vector* subtractVectors(Vector &v);
 	void show();
